Test isolated.
